fixing lint issues

+++ pkg/controller/nodes/task/transformer.go
@@ -100,24 +100,22 @@ func ToTaskExecutionEvent(input ToTaskExecutionEventInputs) (*event.TaskExecutio
 	}
 
 	metadata := &event.TaskExecutionMetadata{
-		GeneratedName:     input.TaskExecContext.TaskExecutionMetadata().GetTaskExecutionID().GetGeneratedName(),
-		PluginIdentifier:  input.PluginID,
-		ResourcePoolInfo:  input.ResourcePoolInfo,
+		GeneratedName:    input.TaskExecContext.TaskExecutionMetadata().GetTaskExecutionID().GetGeneratedName(),
+		PluginIdentifier: input.PluginID,
+		ResourcePoolInfo: input.ResourcePoolInfo,
 	}
 
 	externalResources := input.Info.Info().ExternalResources
 	if externalResources != nil {
 		metadata.ExternalResources = make([]*event.ExternalResourceInfo, len(externalResources))
 		for idx, e := range input.Info.Info().ExternalResources {
-			metadata.ExternalResources[idx] = &event.ExternalResourceInfo {
+			metadata.ExternalResources[idx] = &event.ExternalResourceInfo{
 				ExternalId:   e.ExternalID,
-				RetryAttempt: uint32(e.RetryAttempt),
+				RetryAttempt: e.RetryAttempt,
 				Phase:        ToTaskEventPhase(e.Phase),
 			}
 		}
 	}
-
-	// TODO hamersaw - populate metadata.InstanceClass?
 
 	tev := &event.TaskExecutionEvent{
 		TaskId:                taskExecID.TaskId,
